Reorganize utils and QueryDataUtils
