+++ src/pages/QueryData/QueryDataUtils.jsx
@@ -49,29 +49,4 @@ export const tableToWidget = (table, columns) => ({
   ),
 });
 
-const isIsoDate = str => {
-  if (!/\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z/.test(str)) return false;
-  const d = new Date(str);
-  return d instanceof Date && !Number.isNaN(d) && d.toISOString() === str; // valid date
-};
-
-const dateOptions = {
-  weekday: 'long',
-  year: 'numeric',
-  month: 'long',
-  day: 'numeric',
-  timeZone: 'UTC',
-};
-
-export const humanizeCell = (text, columnType) => {
-  if (isIsoDate(text)) {
-    // (TODO andrew): fix eventually: js time is funky
-    return new Date(text).toLocaleDateString(undefined, dateOptions);
-  }
-  if (columnType === 'tinyint') {
-    return Boolean(text).toString();
-  }
-  return text;
-};
-
-export const TABLE_PRIMARY_KEYS = ['sid', 'rid', 'cid'];
+export const QUERY_PRIMARY_KEYS = ['sid', 'rid', 'cid'];
